requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
matplotlib
seaborn
wordcloud
tqdm
requests
pydantic
